third_party/WebKit/LayoutTests/http/tests/security/img-crossorigin-redirect-credentials-expected.txt

CONSOLE ERROR: Image from origin 'http://localhost:8000' has been blocked from loading by Cross-Origin Resource Sharing policy: No 'Access-Control-Allow-Origin' header is present on the requested resource. Origin 'http://127.0.0.1:8000' is therefore not allowed access.
CONSOLE ERROR: Image from origin 'http://127.0.0.1:8080' has been blocked from loading by Cross-Origin Resource Sharing policy: The 'Access-Control-Allow-Origin' header has a value 'http://127.0.0.1:8000' that is not equal to the supplied origin. Origin 'null' is therefore not allowed access.
CONSOLE ERROR: Image from origin 'http://localhost:8000' has been blocked from loading by Cross-Origin Resource Sharing policy: A wildcard '*' cannot be used in the 'Access-Control-Allow-Origin' header when the credentials flag is true. Origin 'http://127.0.0.1:8000' is therefore not allowed access.
CONSOLE ERROR: Image from origin 'http://127.0.0.1:8000' has been blocked from loading by Cross-Origin Resource Sharing policy: A wildcard '*' cannot be used in the 'Access-Control-Allow-Origin' header when the credentials flag is true. Origin 'null' is therefore not allowed access.
Testing the handling of CORS-enabled fetch in the presence of 'credentialled' redirects.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS/FAIL descriptions are of the form, 'CORS request type': 'redirect CORS type' => 'resource'

PASS: Anonymous request: credentialled => no-CORS image resource.
PASS: Anonymous request: credentialled => anonymous CORS image resource (same origin.)
PASS: Anonymous request: credentialled => anonymous CORS image resource (cross origin.)
PASS: Credentialled request: credentialled => credentialled-CORS image resource (same origin.)
PASS: Credentialled request: credentialled => credentialled-CORS image resource (cross origin.)
PASS: Credentialled request: credentialled => anonymous-CORS image resource (same origin.)
PASS: Credentialled request: credentialled => anonymous-CORS image resource (cross origin.)
PASS successfullyParsed is true

TEST COMPLETE
